IRAF PCIX Version 2.15.1a
 * 64-bit platform support

+++ unix/hlib/mkiraf.csh
@@ -28,16 +28,47 @@ if ("`pwd | grep $irafdir`" != "") then
     endif
 endif
 
+
+# Process command-line arguments.
+set 	user_term	= "none"
+set	init		= 0
+set	quiet		= 0
+
+while ($#argv >= 1) 
+   if ("$argv[1]" == "-t" || "$argv[1]" == "-term") then
+        set user_term   = $argv[2]
+	shift
+   else if ("$argv[1]" == "-i" || "$argv[1]" == "-init") then
+        set init	= 1
+   else if ("$argv[1]" == "-q" || "$argv[1]" == "-quiet") then
+        set quiet	= 1
+   else 
+        echo "Unknown flag '"$argv[1]"'"
+   endif
+   shift
+end
+
+
+
 # Make an empty "uparm" (user parameter) directory.
 if (! -e uparm) then
-    echo '-- creating a new uparm directory'
+    if ($quiet == 0) then
+        echo '-- creating a new uparm directory'
+    endif
     mkdir uparm
 else
-    echo -n 'Initialize uparm? (y|n): '
-    set yesno = $<
-    if ($yesno == 'y' || $yesno == 'yes') then
-	echo '-- initializing uparm'
-	rm -rf uparm; mkdir uparm
+    if ($init == 0) then
+        echo -n 'Initialize uparm? (y|n): '
+        set yesno = $<
+        if ($yesno == 'y' || $yesno == 'yes') then
+	    echo '-- initializing uparm'
+	    rm -rf uparm; mkdir uparm
+        endif
+    else
+        if ($quiet == 0) then
+	    echo '-- initializing uparm'
+        endif
+	/bin/rm -rf uparm; mkdir uparm
     endif
 endif
 
@@ -49,14 +80,16 @@ endif
 # Edit the login.cl file, setting the user's home directory, default image
 # directory, and terminal.
 
+if ($user_term == "none") then
+    echo $ttymsg
+    echo -n 'Enter terminal type: '
+    echo $<	| sed -e "s;.*;s+U_TERM+&+;"		>  _sed
+else
+    echo $user_term | sed -e "s;.*;s+U_TERM+&+;"	>  _sed
+endif
 
-echo $ttymsg
-echo -n 'Enter terminal type: '
-
-
-echo $<	| sed -e "s;.*;s+U_TERM+&+;"		>  _sed
-pwd	| sed -e "s;.*;s+U_HOME+&/+;"		>> _sed
-pwd	| sed -e "s;.*;s+U_UPARM+&/uparm/+;"	>> _sed
+pwd	| sed -e "s;.*;s+U_HOME+&/+;"			>> _sed
+pwd	| sed -e "s;.*;s+U_UPARM+&/uparm/+;"		>> _sed
 if (! (-e "$imdir" && -w "$imdir") ) then
     set imdir = HDR$
     whoami	| sed -e "s;.*;s+U_IMDIR+${imdir}/+;"	>> _sed
@@ -68,5 +101,8 @@ whoami	| sed -e "s;.*;s+U_USER+&+;"		>> _sed
 
 sed -f _sed < $iraf/unix/hlib/login.cl > login.cl; rm _sed
 
-echo 'A new LOGIN.CL file has been created in the current directory.'
-echo 'You may wish to review and edit this file to change the defaults.'
+        
+if ($quiet == 0) then
+    echo 'A new LOGIN.CL file has been created in the current directory.'
+    echo 'You may wish to review and edit this file to change the defaults.'
+endif
